refactor(server): extract content population into documents/populate
ADR-034 Unit 7a, part 2 of 2. Part 1 (#1643) made the mcp/ <-> documents/
boundary observable; this begins the move it was built to police.

`prepareContent` / `applyPreparedContent` / `clearDocMaps` and the two
loaders are the half the open pipeline and the reload family genuinely
share: `clearAndReload` and `reloadFromDisk` reach for exactly what a
first-time open does. Moving the pipeline while leaving them behind would
have made `documents/` import back into `mcp/` -- the cycle 7a exists to
remove, pointed the other way. So the shared half moves first, into a
module that touches neither the registry nor document-service, which is
what lets both sides import it.

The boundary guard caught all three consequences of the move rather than
absorbing them: the new fan-in edge, populate.ts's eight fan-out edges,
and the test-only `__testEvictPartialDocState` alias leaving
file-opener's runtime export surface. Its tables are updated from source,
not from the plan.
